refact :: [#50] Login -> Signin, AuthDomain Interface 파일 구조 정리
