+++ Projects/App/Sources/Application/DI/Authentication/AppComponent+Authentication.swift
@@ -32,9 +32,9 @@ public extension AppComponent {
         }
     }
 
-    var loginUseCase: any LoginUseCase {
+    var signinUseCase: any SigninUseCase {
         shared {
-            LoginUseCaseImpl(repository: authRepository)
+            SigninUseCaseImpl(repository: authRepository)
         }
     }
