+++ Projects/Domain/AuthDomain/Interface/DTO/TokenDTO.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public struct TokenDTO: Codable {
+    public let accessToken: String
+    public let refreshToken: String
+
+    enum CodingKeys: String, CodingKey {
+        case accessToken = "access_token"
+        case refreshToken = "refresh_token"
+    }
+}
