+++ Projects/Domain/AuthDomain/Interface/DataSource/LocalAuthDataSource.swift
@@ -1,0 +1,5 @@
+import Combine
+
+public protocol LocalAuthDataSource {
+    func logout()
+}
