+++ Projects/Domain/AuthDomain/Interface/DataSource/RemoteAuthDataSource.swift
@@ -1,0 +1,6 @@
+import Combine
+
+public protocol RemoteAuthDataSource {
+    func signin(req: SigninRequestParams) -> AnyPublisher<TokenDTO, Error>
+    func refreshToken() -> AnyPublisher<TokenDTO, Error>
+}
