+++ Projects/Domain/AuthDomain/Interface/Entity/TokenEntity.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public struct TokenEntity: Equatable {
+    public let accessToken: String
+    public let refreshToken: String
+    public init(accessToken: String, refreshToken: String) {
+        self.accessToken = accessToken
+        self.refreshToken = refreshToken
+    }
+}
