+++ Projects/Domain/AuthDomain/Interface/Error/AuthError.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+public enum AuthError: Error {
+    case idMismatch
+    case passwordMismatch
+}
+
+extension AuthError: LocalizedError {
+    public var errorDescription: String? {
+        switch self {
+        case .idMismatch:
+            return "올바른 아이디를 입력해주세요"
+        case .passwordMismatch:
+            return "올바른 비밀번호를 입력해주세요"
+        }
+    }
+}
