+++ Projects/Domain/AuthDomain/Interface/Error/PiCKError.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+public enum PiCKError: Error {
+    case error(message: String = "에러가 발생했습니다.", errorBody: [String: Any] = [:])
+    case serverError
+}
+
+extension PiCKError: LocalizedError {
+    public var errorDescription: String? {
+        switch self {
+        case let .error(message, _):
+            return message
+        case .serverError:
+            return "서버 에러가 발생했습니다."
+        }
+    }
+}
