+++ Projects/Domain/AuthDomain/Interface/Repository/AuthRepository.swift
@@ -1,0 +1,7 @@
+import Combine
+
+public protocol AuthRepository {
+    func signin(req: SigninRequestParams) -> AnyPublisher<Void, Error>
+    func refreshToken() -> AnyPublisher<Void, Error>
+    func logout()
+}
