+++ Projects/Domain/AuthDomain/Interface/UseCase/RefreshUseCase.swift
@@ -1,0 +1,5 @@
+import Combine
+
+public protocol RefreshTokenUseCase {
+    func execute() -> AnyPublisher<Void, Error>
+}
