+++ Projects/Domain/AuthDomain/Interface/UseCase/SigninUseCase.swift
@@ -1,0 +1,5 @@
+import Combine
+
+public protocol SigninUseCase {
+    func execute(req: SigninRequestParams) -> AnyPublisher<Void, Error>
+}
