+++ Projects/Domain/AuthDomain/Sources/Repository/AuthRepositoryImpl.swift
@@ -14,8 +14,8 @@ public class AuthRepositoryImpl: AuthRepository {
         self.remoteDataSource = remoteDataSource
     }
 
-    public func login(req: LoginRequestParams) -> AnyPublisher<Void, Error> {
-        remoteDataSource.login(req: req)
+    public func signin(req: SigninRequestParams) -> AnyPublisher<Void, Error> {
+        remoteDataSource.signin(req: req)
             .handleEvents(receiveOutput: { [weak self] tokenData in
                 guard let self = self else { return }
                 self.keyChain.save(type: .accessToken, value: tokenData.accessToken)
