+++ Projects/Domain/AuthDomain/Sources/UaseCase/Auth/SigninUseCaseImpl.swift
@@ -2,14 +2,14 @@ import Foundation
 import Combine
 import AuthDomainInterface
 
-public class LoginUseCaseImpl: LoginUseCase {
+public class SigninUseCaseImpl: SigninUseCase {
     let repository: AuthRepository
 
     public init(repository: AuthRepository) {
         self.repository = repository
     }
 
-    public func execute(req: LoginRequestParams) -> AnyPublisher<Void, Error> {
-        repository.login(req: req)
+    public func execute(req: SigninRequestParams) -> AnyPublisher<Void, Error> {
+        repository.signin(req: req)
     }
 }
